update block class
added initial get,set
